refactor: Coupon/useNewCoupon 생성
