+++ src/refactoring/hooks/index.ts
@@ -1,4 +1,6 @@
-export * from './cart/useCart.ts';
 export * from './useCoupon.ts';
 export * from './useProduct.ts';
+export * from './cart/useCart.ts';
 export * from './admin/useAdmin.ts';
+export * from './admin/useNewCoupon.ts';
+export * from './admin/useNewProduct.ts';
